+++ packages/argdown-parser/src/ArgdownLexer.js
@@ -147,7 +147,8 @@ class ArgdownLexer {
             name: "IncomingSupport",
             pattern: matchIncomingSupport,
             line_breaks: true,
-            label: "+> (Incoming Support)"
+            label: "+> (Incoming Support)",
+            start_chars_hint: [" ", "\t", "+"]
         });
         $.tokens.push($.IncomingSupport);
 
@@ -155,7 +156,8 @@ class ArgdownLexer {
             name: "IncomingAttack",
             pattern: matchIncomingAttack,
             line_breaks: true,
-            label: "-> (Incoming Attack)"
+            label: "-> (Incoming Attack)",
+            start_chars_hint: [" ", "\t", "-"]
         });
         $.tokens.push($.IncomingAttack);
 
@@ -163,7 +165,8 @@ class ArgdownLexer {
             name: "OutgoingSupport",
             pattern: matchOutgoingSupport,
             line_breaks: true,
-            label: "<+ (Outgoing Support)"
+            label: "<+ (Outgoing Support)",
+            start_chars_hint: [" ", "\t", "<"]
         });
         $.tokens.push($.OutgoingSupport);
 
@@ -171,7 +174,8 @@ class ArgdownLexer {
             name: "OutgoingAttack",
             pattern: matchOutgoingAttack,
             line_breaks: true,
-            label: "<- (Outgoing Attack)"
+            label: "<- (Outgoing Attack)",
+            start_chars_hint: [" ", "\t", "<"]
         });
         $.tokens.push($.OutgoingAttack);
 
@@ -179,21 +183,24 @@ class ArgdownLexer {
             name: "Contradiction",
             pattern: matchContradiction,
             line_breaks: true,
-            label: ">< (Contradiction)"
+            label: ">< (Contradiction)",
+            start_chars_hint: [" ", "\t", ">"]
         });
         $.tokens.push($.Contradiction);
         $.IncomingUndercut = createToken({
             name: "IncomingUndercut",
             pattern: matchIncomingUndercut,
             line_breaks: true,
-            label: "_> (Incoming Undercut)"
+            label: "_> (Incoming Undercut)",
+            start_chars_hint: [" ", "\t", "_"]
         });
         $.tokens.push($.IncomingUndercut);
         $.OutgoingUndercut = createToken({
             name: "OutgoingUndercut",
             pattern: matchOutgoingUndercut,
             line_breaks: true,
-            label: "<_ (Outgoing Undercut)"
+            label: "<_ (Outgoing Undercut)",
+            start_chars_hint: [" ", "\t", "<"]
         });
         $.tokens.push($.OutgoingUndercut);
 
@@ -216,7 +223,8 @@ class ArgdownLexer {
             pattern: matchInferenceStart,
             push_mode: "inference_mode",
             line_breaks: true,
-            label: "-- (Inference Start)"
+            label: "-- (Inference Start)",
+            start_chars_hint: [" ", "\t", "-"]
         });
         $.tokens.push($.InferenceStart);
 
@@ -282,7 +290,8 @@ class ArgdownLexer {
             name: "OrderedListItem",
             pattern: matchOrderedListItem,
             line_breaks: true,
-            label: "{Indentation}{number}. (Ordered List Item)"
+            label: "{Indentation}{number}. (Ordered List Item)",
+            start_chars_hint: [" ", "\t"]
         });
         $.tokens.push($.OrderedListItem);
         //whitespace + * + whitespace (to distinguish list items from bold and italic ranges)
@@ -293,7 +302,8 @@ class ArgdownLexer {
             name: "UnorderedListItem",
             pattern: matchUnorderedListItem,
             line_breaks: true,
-            label: "{Indentation}* (Unordered List Item)"
+            label: "{Indentation}* (Unordered List Item)",
+            start_chars_hint: [" ", "\t"]
         });
         $.tokens.push($.UnorderedListItem);
 
@@ -318,7 +328,8 @@ class ArgdownLexer {
             name: "Emptyline",
             pattern: matchEmptyline,
             line_breaks: true,
-            label: "{linebreak}{linebreak} (Empty Line)"
+            label: "{linebreak}{linebreak} (Empty Line)",
+            start_chars_hint: ["\r", "\n"]
         });
         $.tokens.push($.Emptyline);
 
@@ -374,7 +385,7 @@ class ArgdownLexer {
         // });
         // $.tokens.push($.StatementMentionByNumber);
 
-        const statementNumberPattern = /^(?:\r\n|\n|\r)?[' '\t]*\(\d+\)/;
+        const statementNumberPattern = /^[' '\t]*\(\d+\)/;
         function matchStatementNumber(text, offset, matchedTokens, groups) {
             let remainingText = text.substr(offset);
             var last = _.last(matchedTokens);
@@ -395,7 +406,8 @@ class ArgdownLexer {
             name: "StatementNumber",
             pattern: matchStatementNumber,
             line_breaks: true,
-            label: "(Number) (Statement Number)"
+            label: "(Number) (Statement Number)",
+            start_chars_hint: [" ", "\t", "("]
         });
         $.tokens.push($.StatementNumber);
 
@@ -434,7 +446,8 @@ class ArgdownLexer {
         $.HeadingStart = createToken({
             name: "HeadingStart",
             pattern: matchHeadingStart,
-            label: "# (Heading Start)"
+            label: "# (Heading Start)",
+            start_chars_hint: ["#"]
         });
         $.tokens.push($.HeadingStart);
 
@@ -501,56 +514,64 @@ class ArgdownLexer {
         $.AsteriskBoldStart = createToken({
             name: "AsteriskBoldStart",
             pattern: matchAsteriskBoldStart,
-            label: "** (Bold Start)"
+            label: "** (Bold Start)",
+            start_chars_hint: ["*"]
         });
         $.tokens.push($.AsteriskBoldStart);
 
         $.AsteriskBoldEnd = createToken({
             name: "AsteriskBoldEnd",
             pattern: matchAsteriskBoldEnd,
-            label: "** (Bold End)"
+            label: "** (Bold End)",
+            start_chars_hint: ["*"]
         });
         $.tokens.push($.AsteriskBoldEnd);
 
         $.UnderscoreBoldStart = createToken({
             name: "UnderscoreBoldStart",
             pattern: matchUnderscoreBoldStart,
-            label: "__ (Bold Start)"
+            label: "__ (Bold Start)",
+            start_chars_hint: ["_"]
         });
         $.tokens.push($.UnderscoreBoldStart);
 
         $.UnderscoreBoldEnd = createToken({
             name: "UnderscoreBoldEnd",
             pattern: matchUnderscoreBoldEnd,
-            label: "__ (Bold End)"
+            label: "__ (Bold End)",
+            start_chars_hint: ["_"]
         });
         $.tokens.push($.UnderscoreBoldEnd);
 
         $.AsteriskItalicStart = createToken({
             name: "AsteriskItalicStart",
             pattern: matchAsteriskItalicStart,
-            label: "* (Italic Start)"
+            label: "* (Italic Start)",
+            start_chars_hint: ["*"]
         });
         $.tokens.push($.AsteriskItalicStart);
 
         $.AsteriskItalicEnd = createToken({
             name: "AsteriskItalicEnd",
             pattern: matchAsteriskItalicEnd,
-            label: "* (Italic End)"
+            label: "* (Italic End)",
+            start_chars_hint: ["*"]
         });
         $.tokens.push($.AsteriskItalicEnd);
 
         $.UnderscoreItalicStart = createToken({
             name: "UnderscoreItalicStart",
             pattern: matchUnderscoreItalicStart,
-            label: "_ (Italic Start)"
+            label: "_ (Italic Start)",
+            start_chars_hint: ["_"]
         });
         $.tokens.push($.UnderscoreItalicStart);
 
         $.UnderscoreItalicEnd = createToken({
             name: "UnderscoreItalicEnd",
             pattern: matchUnderscoreItalicEnd,
-            label: "_ (Italic End)"
+            label: "_ (Italic End)",
+            start_chars_hint: ["_"]
         });
         $.tokens.push($.UnderscoreItalicEnd);
 
@@ -739,5 +760,5 @@ class ArgdownLexer {
 }
 
 module.exports = {
-    ArgdownLexer: new ArgdownLexer()
+    ArgdownLexer: new ArgdownLexer({ensureOptimizations : true})
 };